+++ Scripts/main.js
@@ -1,14 +1,17 @@
 let disposeTask;
+let sidebarTreeView;
+let sidebarDataProvider;
 
 // Commands
-const { runHTTP } = require('./run-http');
+const { runHTTP, emitter } = require('./run-http');
+const { sidebarData, sidebarTree, clearCmd, addCmd } = require('./sidebar');
 
 nova.commands.register('runHTTP', runHTTP);
 
 // Tasks
 class RunHTTPTask {
   provideTasks() {
-    let task = new Task("Run HTTP");
+    let task = new Task("Run current HTTP");
     task.setAction(Task.Run, new TaskCommandAction('runHTTP', {
       args: []
     }));
@@ -17,10 +20,17 @@ class RunHTTPTask {
 }
 
 disposeTask = nova.assistants.registerTaskAssistant(new RunHTTPTask(), {
-  name: "Run HTTP Assistant"
+  name: "HTTP Client"
 });
 
+nova.commands.register("sidebar.clear", clearCmd);
+
 exports.activate = function() {
+  emitter.on("finished", (req) => {
+    addCmd(req);
+  })  
+  
+  nova.subscriptions.add(sidebarTree);
   console.log('activated')
 }